[DEVOPS-391] Use sequelize for seeds
NB: `yarn seed` isn't actually used. Seeds are currently *internal* to the tests.
